Fixes search bar clobbering
A value selected in the search bar must also be in the options
list, otherwise it's removed.

This was introducing a bug because longer repo addresses were
being dropped but org addresses weren't.

This was because for short inputs of only a few letters, the addresses
of repos were of index >250 so they were dropped when we limited
the length of the returnable options, which had the side-effect of removing
them from the list.

This change ensures that those values currently selected are always appended
to the end of the options list so they won't be removed randomly.

Signed-off-by: JamesKunstle <jkunstle@bu.edu>